Финальная версия проекта

+++ blog/views.py
@@ -3,9 +3,9 @@ from django.utils import timezone
 from django.core.paginator import Paginator
 from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
+from django.http import Http404
 from .models import Post, Category, Comment
 from .forms import CommentForm, PostForm
-from django.shortcuts import redirect
 
 User = get_user_model()
 
@@ -31,23 +31,23 @@ def index(request):
 
 def post_detail(request, id):
     post = get_object_or_404(
-        Post.objects.select_related(
-            'author', 'category', 'location'
-        ),
-        id=id,
-        is_published=True,
-        pub_date__lte=timezone.now(),
-        category__is_published=True
+        Post.objects.select_related('author', 'category', 'location'),
+        id=id
     )
+    
+    # Если пользователь не автор, проверяем условия публикации
+    if request.user != post.author:
+        if not post.is_published or not post.category.is_published or post.pub_date > timezone.now():
+            raise Http404()
+    
     comments = post.comments.all()
     form = CommentForm()
 
-    context = {
+    return render(request, 'blog/detail.html', {
         'post': post,
         'comments': comments,
         'form': form,
-    }
-    return render(request, 'blog/detail.html', context)
+    })
 
 
 def category_posts(request, category_slug):
@@ -79,7 +79,6 @@ def category_posts(request, category_slug):
 def profile(request, username):
     user = get_object_or_404(User, username=username)
     
-    # Для автора показываем все посты, для других - только опубликованные
     if request.user == user:
         posts = Post.objects.filter(author=user).order_by('-pub_date')
     else:
@@ -98,6 +97,7 @@ def profile(request, username):
         'page_obj': page_obj,
     }
     return render(request, 'blog/profile.html', context)
+
 
 @login_required
 def add_comment(request, post_id):
@@ -132,8 +132,10 @@ def edit_comment(request, post_id, comment_id):
 def delete_comment(request, post_id, comment_id):
     comment = get_object_or_404(Comment, id=comment_id, post_id=post_id)
     if comment.author == request.user:
-        comment.delete()
-    return redirect('blog:post_detail', id=post_id)
+        if request.method == 'POST':
+            comment.delete()
+            return redirect('blog:post_detail', id=post_id)
+    return render(request, 'blog/comment.html', {'comment': comment})
 
 
 @login_required
@@ -143,7 +145,7 @@ def create_post(request):
         if form.is_valid():
             post = form.save(commit=False)
             post.author = request.user
-            post.pub_date = timezone.now()  # автоматически текущее время
+            post.pub_date = timezone.now()
             post.save()
             return redirect('blog:profile', username=request.user.username)
     else:
@@ -177,47 +179,4 @@ def delete_post(request, post_id):
         post.delete()
         return redirect('blog:profile', username=request.user.username)
 
-    return render(request, 'blog/create.html', {'form': post})
-
-@login_required
-def create_post(request):
-    if request.method == 'POST':
-        form = PostForm(request.POST, request.FILES)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            post.save()
-            return redirect('blog:profile', username=request.user.username)
-    else:
-        form = PostForm()
-    return render(request, 'blog/create.html', {'form': form})
-@login_required
-def edit_post(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    
-    # Проверка прав: только автор может редактировать
-    if post.author != request.user:
-        return redirect('blog:post_detail', id=post_id)
-    
-    if request.method == 'POST':
-        form = PostForm(request.POST, request.FILES, instance=post)
-        if form.is_valid():
-            form.save()
-            return redirect('blog:post_detail', id=post_id)
-    else:
-        form = PostForm(instance=post)
-    
-    return render(request, 'blog/create.html', {'form': form})
-@login_required
-def delete_post(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    
-    # Проверка прав: только автор может удалить
-    if post.author != request.user:
-        return redirect('blog:post_detail', id=post_id)
-    
-    if request.method == 'POST':
-        post.delete()
-        return redirect('blog:profile', username=request.user.username)
-    
     return render(request, 'blog/delete.html', {'post': post})
